[stats] added fitting statistics and avoid splitting components with kappa close to MAX_KAPPA
Avoids splitting a component with a kappa close to MAX_KAPPA. The splitting metric wants to split these components if the signal is narrower then a VMF with kappa of MAX_KAPPA. Since the fitting blocks higher kappas this would lead to a recursive loop of split and merge.

Added statistics of the number of performed split and merges as well as the number of EM iterations for the update and partial-update fitting steps.